wip: the last refactor
