+++ core/src/text/fontContext.h
@@ -1,5 +1,11 @@
 #pragma once
 
+#include "textUtil.h"
+
+// For textParameters
+#include "style/textStyle.h"
+#include "labels/textLabel.h"
+
 #include "alfons/alfons.h"
 #include "alfons/fontManager.h"
 #include "alfons/atlas.h"
@@ -15,9 +21,6 @@
 
 namespace Tangram {
 
-constexpr int textureSize = 256;
-constexpr int maxTextures = 64;
-
 namespace alf = alfons;
 
 struct FontMetrics {
@@ -27,8 +30,10 @@ struct FontMetrics {
 // TODO could be a shared_ptr<Texture>
 struct GlyphTexture {
 
-    GlyphTexture() : texture(textureSize, textureSize) {
-        texData.resize(textureSize * textureSize);
+    static constexpr int size = 256;
+
+    GlyphTexture() : texture(size, size) {
+        texData.resize(size * size);
     }
 
     std::vector<unsigned char> texData;
@@ -39,7 +44,11 @@ struct GlyphTexture {
 };
 
 class FontContext : public alf::TextureCallback {
+
 public:
+
+    static constexpr int max_textures = 64;
+
     FontContext();
 
     /* Synchronized on m_mutex on tile-worker threads
@@ -55,13 +64,11 @@ public:
     void addGlyph(alf::AtlasID id, uint16_t gx, uint16_t gy, uint16_t gw, uint16_t gh,
                   const unsigned char* src, uint16_t pad) override;
 
-    void releaseAtlas(std::bitset<maxTextures> _refs);
+    void releaseAtlas(std::bitset<max_textures> _refs);
 
-    void lockAtlas(std::bitset<maxTextures> _refs);
+    void lockAtlas(std::bitset<max_textures> _refs);
 
     alf::GlyphAtlas& atlas() { return m_atlas; }
-
-    std::mutex& mutex() { return m_mutex; }
 
     /* Update all textures batches, uploads the data to the GPU */
     void updateTextures();
@@ -77,17 +84,37 @@ public:
 
     float maxStrokeWidth() { return m_sdfRadius; }
 
+    bool layoutText(TextStyle::Parameters& _params, Label::Type _type,
+                    const std::string& _text,
+                    std::vector<GlyphQuad>& _quads, glm::vec2& _bbox);
+
+    struct ScratchBuffer : public alfons::MeshCallback {
+        void drawGlyph(const alfons::Quad& q, const alfons::AtlasGlyph& altasGlyph) override {}
+        void drawGlyph(const alfons::Rect& q, const alfons::AtlasGlyph& atlasGlyph) override;
+        std::vector<GlyphQuad>* quads;
+    };
+
 private:
     float m_sdfRadius;
+    ScratchBuffer m_scratch;
 
     std::mutex m_mutex;
-    std::array<int, maxTextures> m_atlasRefCount = {{0}};
+    std::array<int, max_textures> m_atlasRefCount = {{0}};
     alf::GlyphAtlas m_atlas;
 
     alf::FontManager m_alfons;
     std::array<std::shared_ptr<alf::Font>, 3> m_font;
 
     std::vector<GlyphTexture> m_textures;
+
+    // TextShaper to create <LineLayout> for a given text and Font
+    alfons::TextShaper m_shaper;
+
+    // TextBatch to 'draw' <LineLayout>s, i.e. creating glyph textures and glyph quads.
+    // It is intialized with a TextureCallback implemented by FontContext for adding glyph
+    // textures and a MeshCallback implemented by TextStyleBuilder for adding glyph quads.
+    alfons::TextBatch m_batch;
+
 };
 
 }
